+++ spec/bedrock_runtime/response_builders/ai_21_labs_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-require 'bedrock_runtime/response_builders/text'
+require 'bedrock_runtime/response_builders/ai_21_labs'
 
-RSpec.describe RubyAmazonBedrock::ResponseBuilders::Text do
+RSpec.describe RubyAmazonBedrock::ResponseBuilders::Ai21Labs do
   let(:response_body) { { key1: 'value1', key2: 'value2' }.to_json }
   let(:response) { double('response', body: StringIO.new(response_body)) }
 
@@ -14,8 +14,7 @@ RSpec.describe RubyAmazonBedrock::ResponseBuilders::Text do
       result = subject.build
       expect(result).to be_a(Hash)
       expect(result.keys).to all(be_a(Symbol))
-      expect(result[:key1]).to eq('value1')
-      expect(result[:key2]).to eq('value2')
+      expect(result[:full_response]).to eq(JSON.parse(response_body, symbolize_names: true))
     end
   end